Add support for FileAppender file reopening
This is part of a series of patches to support log rotation.

A function is added to the `Append` trait which is called after log
rotation. Per earlier discussion with @sfackler, this is not an ideal
API since windows doesn't let you rename open files. There does need to
be some sort of interface on Append to accomplish this, though, since
the SharedLogger keeps a map of `Append` types.

There is a bit of logic repeated between post_rotate and
`FileAppenderBuilder::build` as far as opening the file and creating the
`SimpleWriter`. This should be refactored in the future.

The `post_rotate` default implementation is a noop. The FileAppender
implementation obtains a lock on current file, flushes any pending
writes, attempts to open the new file, and finally, if that succeeds,
swaps the new writer with the old. The old writer falls out of scope and
closes the underlying file handle, and the new writer is used for
subsequent log operations.